move composition policy check to plugin
